+++ tests/Unit/Wye/SetBacktraceDefaultLimitTest.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Tests\Unit\Wye;
+
+use Stratedge\Wye\Wye;
+
+class SetBacktraceDefaultLimitTest extends \Tests\TestCase
+{
+    public function testSetsDefaultLimit()
+    {
+        Wye::setBacktraceDefaultLimit(19);
+
+        $this->assertSame(19, Wye::getBacktraceDefaultLimit());
+    }
+}